pyproject.toml
[project]
name = "voice-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-sdk-bedrock-runtime>=0.1.0",
    "boto3[crt]>=1.42.55",
    "fastapi>=0.133.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
    "strands-agents[bidi]>=1.27.0",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.21",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
    "chromadb>=1.5.1",
    "langchain-text-splitters>=1.1.1",
    "pymupdf>=1.27.1",
    "nest-asyncio>=1.6.0",
    "httpx>=0.28.1",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.10.0",
]
